Correções Instaladores do Integrador e do Monitorador

+++ Instalador/Windows/Arquivos de Programas PC/Instalador_Integrador.bat
@@ -300,12 +300,15 @@ if %tipoInstalacao% == 1 (
 
 cd "Arquivos de Programas PC"
 
-del /f /q *.reg
-del /f /q *.exe
+REM Os seguintes arquivos não existem se optar por instalação usando uma JRE já existente.
+REM Para evitar erro indevido "Arquivo não encontrado" usar if exist:
+if exist *.reg del /f /q *.reg
+if exist *.exe del /f /q *.exe
+
 del /f /q Instalador.bat
 del /f /q TestadorUnitarioInstalacaoDesinstalacao.bat
-del /f /q UpgradeNaoNuvem3.1.0.bat
-del /f /q .gitignore
+if exist UpgradeNaoNuvem3.1.0.bat del /f /q UpgradeNaoNuvem3.1.0.bat
+if exist .gitignore del /f /q .gitignore
 
 if %tipoInstalacao% == 1 (
     del /f /q AdicionarFornecedorNuvem.bat
@@ -401,10 +404,12 @@ if %tipoInstalacao% == 2 (
     del %temp%\TEMPmessage.vbs /f /q
 ) else if %tipoInstalacao% == 1 (
     echo.
-    echo          Primeira fase da instalação concluida!
+    echo          Primeira fase da instalação concluida! 
+    echo          Favor preencher o arquivo de configuração, 
+    echo          já aberto automaticamente em segundo plano.
     echo.
     
-    echo MSGBOX "Primeira fase da instalação concluida!" > %temp%\TEMPmessage.vbs
+    echo MSGBOX "Primeira fase da instalação concluida! Favor preencher o arquivo de configuração, já aberto automaticamente em segundo plano." > %temp%\TEMPmessage.vbs
     call %temp%\TEMPmessage.vbs
     del %temp%\TEMPmessage.vbs /f /q
